LLMlogs/requirements_qwen.txt
# Core ML and Data Processing# Core ML and Data Processing

numpy>=1.23.5,<2.3numpy>=1.23.5,<2.3

pandas>=2.0.0pandas>=2.0.0

scikit-learn>=1.3.0scikit-learn>=1.3.0

xgboost>=2.0.0xgboost>=2.0.0

prophet>=1.1.0prophet>=1.1.0

joblib>=1.3.0joblib>=1.3.0

scipy>=1.10.0scipy>=1.10.0



# HTTP and API# HTTP and API

requests>=2.28.0requests>=2.28.0

python-dotenv>=1.0.0python-dotenv>=1.0.0



# Kafka for log streaming# Kafka for log streaming

kafka-python>=2.0.0kafka-python>=2.0.0



# Cloud providers# Cloud providers

boto3>=1.34.0boto3>=1.34.0

azure-monitor-query>=2.0.0azure-monitor-query>=2.0.0

azure-identity>=1.15.0azure-identity>=1.15.0



# Visualization (optional for dashboards)# Visualization (optional for dashboards)

plotly>=5.0.0plotly>=5.0.0

matplotlib>=3.7.0matplotlib>=3.7.0

seaborn>=0.12.0seaborn>=0.12.0



# LangChain and LLM dependencies for Qwen integration# LangChain and LLM dependencies for Qwen integration

langchain>=0.1.0langchain>=0.1.0

langchain-community>=0.0.1langchain-community>=0.0.1

llama-cpp-python>=0.2.0llama-cpp-python>=0.2.0

huggingface-hub>=0.19.0huggingface-hub>=0.19.0

transformers>=4.36.0transformers>=4.36.0

torch>=2.1.0torch>=2.1.0

accelerate>=0.25.0accelerate>=0.25.0



# Additional utilities for model management# Additional utilities for model management

safetensors>=0.4.0safetensors>=0.4.0

tokenizers>=0.15.0tokenizers>=0.15.0